requirements.txt
celery==5.0.5
retry==0.9.2
Flask==1.1.2
validator.py==1.3.0
requests==2.20.1
pycryptodome==3.10.1
PyJWT==2.0.1
validator.py==1.3.0
pycryptodome==3.10.1
redis==3.5.3
supervisord